move toArgv to own file
