+++ butterfly/__init__.py
@@ -1,5 +1,5 @@
-__all__ = ["main", "Webserver", "Settings"]
-__all__ += ["AccessLayer", "DatabaseLayer"]
+__all__ = ["main", "Webserver", "toArgv"]
+__all__ += ["Settings","AccessLayer", "DatabaseLayer"]
 __all__ += ["CoreLayer", "ImageLayer", "QueryLayer"]
 
 import Settings
@@ -10,3 +10,4 @@ import AccessLayer
 import DatabaseLayer
 from .Webserver import Webserver
 from .butterfly import main
+from .toArgv import toArgv
