+++ butterfly/butterfly.py
@@ -1,4 +1,5 @@
 from Webserver import Webserver
+from toArgv import toArgv
 import sys, argparse
 import logging
 
@@ -18,7 +19,7 @@ class Butterfly():
         port = args['port']
 
         logging.basicConfig(**self.log_info)
-        self._server = Webserver().start(port)
+        Webserver().start(port)
 
     def parseArgv(self, argv):
         sys.argv = argv
@@ -35,16 +36,6 @@ class Butterfly():
         parser.add_argument('-e','--exp', metavar='exp', help= help['folder'])
         parser.add_argument('-o','--out', metavar='out', help= help['save'])
         return vars(parser.parse_args())
-
-# Allow Common Command Line / Module Interface
-
-def toArgv(*args, **flags):
-    keyvals = flags.items()
-    all_tokens = range(2*len(keyvals))
-    endash = lambda fkey: '-'+fkey if len(fkey) == 1 else '--'+fkey
-    enflag = lambda kv,fl: str(kv[fl]) if fl else endash(str(kv[fl]))
-    kargv = [enflag(keyvals[i//2],i%2) for i in all_tokens]
-    return ['main'] + list(map(str,args)) + kargv
 
 def main(*_args, **_flags):
     return Butterfly(toArgv(*_args, **_flags))
