+++ unit_tests/Test.py
@@ -36,18 +36,8 @@ class Test():
         parser.add_argument('unit', nargs='?', default='Webserver', help=help['unit'])
         return vars(parser.parse_args())
 
-# Allow Common Command Line / Module Interface
-
-def toArgv(*args, **flags):
-    keyvals = flags.items()
-    all_tokens = range(2*len(keyvals))
-    endash = lambda fkey: '-'+fkey if len(fkey) == 1 else '--'+fkey
-    enflag = lambda kv,fl: str(kv[fl]) if fl else endash(str(kv[fl]))
-    kargv = [enflag(keyvals[i//2],i%2) for i in all_tokens]
-    return ['main'] + list(map(str,args)) + kargv
-
 def main(*_args, **_flags):
-    return Test(toArgv(*_args, **_flags))
+    return Test(butterfly.toArgv(*_args, **_flags))
 
 if __name__ == "__main__":
     print Test(sys.argv).start()
